L4R/LRR/tools/lucxbox/requirements.txt
altgraph==0.16.1
dohq-artifactory==0.7.630
astroid==2.4.2
atomicwrites>=1.3.0
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
colorama==0.3.9
coverage==4.5.3
flake8==3.8.3
future==0.17.1
idna==2.7
importlib-metadata==0.18
isort==4.3.21
jsonschema==2.6.0
lazy-object-proxy==1.4.1
macholib==1.11
matplotlib==3.3.0
mccabe==0.6.1
more-itertools==4.3.0
pandas==1.3.3
pefile==2019.4.18
pluggy==0.12.0
py==1.8.0
Jinja2==2.10.1
pygments==2.2.0
PyInstaller==4.0
pylint==2.5.3
pytest==3.8.2
pytest-cov==2.6.0
pytest-mock==1.10.0
pywin32-ctypes==0.2.0
PyYAML==5.3.1
requests==2.19.1
six==1.12.0
stashy==0.7
tqdm==4.26.0
urllib3==1.23
wheel==0.33.4
wrapt==1.11.2
XlsxWriter==1.0.5
zipp==0.5.1
GitPython==3.1.1
gitdb==4.0.4
